Consolidate List and ListItem
