+++ src/components/ComponentFactory.js
@@ -37,8 +37,8 @@ import Introduction from './Introduction';
 import Kicker from './Kicker';
 import Line from './Line';
 import LineBlock from './LineBlock';
-import List from './List';
-import ListItem from './ListItem';
+import List from './List/List';
+import ListItem from './List/ListItem';
 import ListTable from './ListTable';
 import Literal from './Literal';
 import LiteralBlock from './LiteralBlock';
